move default classes to theme view helper
